requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch
torchvision
torchattacks>=3.4.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
Pillow>=9.5.0
tqdm>=4.65.0
scipy>=1.11.0
opencv-python-headless>=4.8.0
grad-cam>=1.4.8
kaggle>=1.5.16
pytorch-msssim>=0.2.1
tabulate
